perf(parser): zero-alloc dispatcher Counter cache (Rust 1.95.0 follow-up) (#370)
* chore(toolchain): bump Rust 1.93.1 to 1.95.0

* fix(clippy): address Rust 1.95.0 lib lints across workspace

* fix(clippy): address Rust 1.95.0 test-target lints across workspace

* perf(parser): cache dispatcher Counter handles to keep hot path zero-alloc

The `metrics::counter!()` macro with labels allocates a `Vec<Label>` per
invocation. `dispatch_frame` calls it on every WebSocket frame, which
showed up as 105 heap blocks in `dhat_all_parsers_zero_alloc` (1 alloc
per ticker/quote/depth/full + 100-burst + error path), violating
Principle #1.

Fix: cache the resolved `Counter` handle in a `OnceLock` per response-
code label. The dispatcher hot path now does a static-cell lookup +
`.increment(1)` on the cached handle — no allocation after the first
call. The unused `response_code_label` helper is removed.

A new `prewarm_dispatcher_counters()` is called from app boot right
after `init_metrics()` so production observes zero allocations from the
very first packet (a handle created before the recorder is installed
resolves to a no-op forever, so prewarm MUST run post-install). The
DHAT test calls the same prewarm before the measurement window so the
zero-alloc assertion holds in isolation.

Verification:
- cargo test -p tickvault-core --test dhat_allocation — passes (was 105 blocks)
- cargo test -p tickvault-core --lib — 3216 passed
- cargo clippy -p tickvault-core -p tickvault-app --all-targets -- -D warnings — clean
- cargo fmt --check — clean


---------

+++ crates/trading/benches/obi.rs
@@ -16,7 +16,7 @@ fn make_levels(base_price: f64, step: f64, base_qty: u32) -> Vec<DeepDepthLevel>
         .map(|i| DeepDepthLevel {
             price: base_price + step * i as f64,
             quantity: base_qty.saturating_sub(i * (base_qty / 25)),
-            orders: (20 - i) as u32,
+            orders: 20 - i,
         })
         .collect()
 }
